Revert "Implement PartialRolloutFn abstraction for clean partial rollout (#172)"
This reverts commit 2a95a217c26f15deaedcdd4e0ed10de3a909182d.
