+++ slime/rollout/sglang_rollout.py
@@ -1,6 +1,5 @@
 import asyncio
 import copy
-from functools import partial
 
 from tqdm import tqdm
 from transformers import AutoTokenizer
@@ -11,12 +10,10 @@ from slime.utils.http_utils import get, post
 from slime.utils.misc import SingletonMeta, load_function
 from slime.utils.types import Sample
 from slime.rollout.components.sample_generator import generate_one_sample_vanilla
-from .components.base_rollout_fn import RolloutFnCallParams, RolloutFnInitParams, RolloutFnCallOutput
-from .components.partial_rollout_fn import PartialRolloutFn
 
 from .rm_hub import async_rm, batched_async_rm
 
-__all__ = ["create_rollout_fn"]
+__all__ = ["generate_rollout"]
 
 
 class GenerateState(metaclass=SingletonMeta):
@@ -157,13 +154,13 @@ async def abort(args, rollout_id: int):
     return aborted_samples
 
 
-async def generate_rollout_async(args, rollout_id: int, get_samples):
+async def generate_rollout_async(args, rollout_id: int, data_source) -> list[list[Sample]]:
     """An example to implement the generate_rollout function for an rule based rm rollout generation.
 
     Args:
         args: the whole args
         rollout_id: int, the id of the rollout, used for deterministic data generation
-        get_samples: the data source to fetch
+        data_source: the data source to fetch
 
     Returns:
         list[list[Sample]]: a list of samples generated by the rollout, the length of the list is exactly the same as the `rollout_batch_size`
@@ -189,7 +186,7 @@ async def generate_rollout_async(args, rollout_id: int, get_samples):
     while len(data) < target_data_size:
         while state.remaining_batch_size < target_data_size:
             # get samples from the buffer and submit the generation requests.
-            samples = get_samples(args.over_sampling_batch_size)
+            samples = data_source(args.over_sampling_batch_size)
             state.submit_generate_tasks(samples)
 
         # wait for the generation to finish
@@ -232,7 +229,7 @@ async def generate_rollout_async(args, rollout_id: int, get_samples):
 
     # reset the global state to prevent effects on the next rollout or eval.
     state.reset()
-    return RolloutFnCallOutput(samples=data), aborted_samples
+    return data, aborted_samples
 
 
 EVAL_PROMPT_DATASET = {}
@@ -244,7 +241,7 @@ async def eval_rollout(args, rollout_id):
     for i in range(0, len(args.eval_prompt_data), 2):
         name, path = args.eval_prompt_data[i : i + 2]
         results.update(await eval_rollout_single_dataset(args, rollout_id, name, path))
-    return RolloutFnCallOutput(metrics=results), []
+    return results, []
 
 
 async def eval_rollout_single_dataset(args, rollout_id, name, path):
@@ -329,20 +326,28 @@ async def eval_rollout_single_dataset(args, rollout_id, name, path):
     }
 
 
-def _generate_one_step(
-    init_params: RolloutFnInitParams,
-    params: RolloutFnCallParams,
-    get_samples,
-):
-    if init_params.evaluation:
-        return run(eval_rollout(init_params.args, params.rollout_id))
-    else:
-        return run(generate_rollout_async(init_params.args, params.rollout_id, get_samples))
+# TODO remove this temp function
+def generate_rollout(args, rollout_id, data_buffer, evaluation=False):
+    """An example to implement the generate_rollout function for an rule based rm rollout generation.
 
+    Args:
+        args: the whole args
+        rollout_id: int, the id of the rollout, used for deterministic data generation
+        data_buffer: the data buffer to store the generated samples
+        evaluation: bool, whether the rollout is for evaluation or not
 
-def create_rollout_fn(params: RolloutFnInitParams):
-    assert params.args.rollout_global_dataset
-    return PartialRolloutFn(
-        params=params,
-        generate_one_step=partial(_generate_one_step, init_params=params),
+    Returns:
+        list[list[Sample]]: a list of list of samples generated by the rollout
+    """
+    completed_samples, aborted_samples = generate_abortable_samples(
+        args, rollout_id, data_buffer.get_samples, evaluation=evaluation
     )
+    data_buffer.add_samples(aborted_samples)
+    return completed_samples
+
+
+def generate_abortable_samples(args, rollout_id, data_source, evaluation=False):
+    assert args.rollout_global_dataset
+    if evaluation:
+        return run(eval_rollout(args, rollout_id))
+    return run(generate_rollout_async(args, rollout_id, data_source))
